Fix npm build warnings
This fixes a handful of npm build warnings that were showing up
when we started to build our frontend app for production.

We also add a build_and_push script that will build the frontend
container image and push it up to our registry.

We also update the Dockerfile build to remove the api environment
variable since it's a pointless variable during runtime. The app
will have already been compiled so the environment variable doesn't
affect the API URL during runtime. Instead, we fix the API url to
either be localhost if development mode or api.tincan.com if in
production mode.

Lastly, we update a bunch of node modules to their latest version.