Add hasError prop to subcategory selector
This adds an optional boolean prop to the subcategory selector
component. The prop is hasError, and is a boolean value that
indicates there is an error in the selection and the component
should enable visual aids to show the user that there is an error.

The visual aids are a red border around the entire selector.