Add Transactions Splits UI
This adds a new modal that allows the user to split transactions.
The modal is accessed by clicking "split transaction" in the
transaction expanded drawer.

The modal will show the original transaction, as well as a button
to add splits. The splits are rendered with a new component called
EditableTransactionSplit. We weren't able to reuse the
EditableTransactionRow since they have very little overlap in
functionality.

The validation for splits is handled in the modal parent component,
but we also have redundant validation in the Editable Splits
component to give the user visual aids as to where the errors are
located.

The splits are paginated to show 5 splits per page, otherwise the
modal becomes too long and functionally becomes unusable since
buttons start to get hidden.