Update SubcategorySelector to float on top
This updates the SubcategorySelector UI code to attach it to the
document.body directly instead of the child components. This is
necessary because scrollable divs or any relative div will cause
the dropdown to get cut off on its boundaries. Now, the dropdown
will always show up regardless of the div that it's rendered in.