Refactor Dockerfile to reduce image size
This refactors the frontend Dockerfile to implement build stages as
well as use a slim base image. The final deploy image doesn't need
the devDependencies so we exclude them in the final stage.
Additionally, the build stage doesn't need the production
dependencies so we only include dev dependencies.

We also introduce a dockerignore file to scope down what gets
copied into the image.

With this change we reduced the image size from 3.56GB to 2.21GB.
That's over a gigabyte of savings!