Reafactor components to fit new definitions
This refactors a hand full of transaction components to update the
new data definitions coming from the backend. user name and account
name are no longer passed with the transcation object. They must be
joined from the account and user data.

We also remove the custodian from the formatAccount function. The
custodian is going to be placed in different locations depending on
the component, so the component should be responsible for placing
the custodian rather than appending it to the account name.