Refactor Accounts components to match definititions
This commit refactors a handful of the Account components to match
the newly update type definitions. We convert account and user ids
to strings in an effort to get all IDs to a string. To keep the
commits smaller, we only update the IDs on Accounts and Users, and
the remaining type definitions will be updated in future commits.