Refactor Accounts page
This commit is a major refactor of the Accounts page. The new page
will display each account grouped by the account_type and then
further grouped by the account_subtype. The institution_name,
account name and custodial's user name will be listed in a column
on the left hand side of the account row. The current_balance of
the account will be shown on the right hand side of the row.

In an effort to get this page usable quickly, we'll have to put off
adding click handlers and listeners on the rows which means users
can't edit or delete the accounts for now. This is fine because no
one is using the app at this time and I can easily go into the
Rails console to make edits. However before we launch we'll need a
way to modify and delete accounts.

This also deletes the existing Accounts filters component. I'll
think about if we need filters on this page at all, and if we do
want them, there is a new design for the filters which means this
old component will never get used again.