Refactor AddAccountModal
This is a major refactor of the add account modal. The modal now
first asks the user how they would like to add a new account: link
to bank via Plaid or manual entries.

We also rework the Modal to be a state machine and keep track of
the screen we are displaying. We also move the back button to the
new header and remove the back button from the add details screen.

The manual account functions and variables have been renamed to
make it clear that we are referring to manual accounts, and all
other accounts are Plaid linked accounts.