PlaidLink: Update job statuses to status
This refactors the Plaid Item initialization job status API to be a
singular status rather than two statuses. This is due to the
backend change that consolidated the item initialization logic into
one job: AccountsSyncJob which will now pull the item information
as well as the accounts information.

We also add another success status that appears once the jobs have
completed to let the user know the bank has been successfully
linked, and that transactions will sync within 5 minutes.