Update API URL to use api.localhost
This updates the API URL in the development environment to use
api.localhost which nginx is configured to listen for on the dev
servers. Production will still use api.tincan.com, but we'll
need to update that once the app is open on the public internet.
For now, the app is being used only on a local network which has
manual DNS entries for tincan.com.