ApiUtils: base URL hardcode
This updates the hardcoded API base URL to get us working in
production. We probably want to pass this in as a build time
variable but since I'm in a rush to get this working in production,
lets just hardcode it to the value we know it will be for our
deployment.